backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.12.1
python-decouple==3.8
razorpay==2.0.1
sqlparse==0.5.5
stripe==11.5.0
gunicorn==21.2.0
whitenoise==6.6.0
django-ratelimit==4.1.0
django-jazzmin==3.0.4
psycopg2-binary==2.9.10
django-filter==24.3
